+++ app/controllers/lessons_controller.rb
@@ -1,35 +1,59 @@
 class LessonsController < ApplicationController
   before_action :set_lesson, only: [:show, :edit, :update, :destroy]
-  access all: [:show, :set_lesson],  tutor: [:new, :create, :show, :index, :edit, :update, :destroy, :set_lesson, :lesson_params]
+  access all: [:index, :show, :new, :edit, :create, :update, :destroy], user: :all
 
+  # GET /lessons
+  def index
+    @lessons = Lesson.all
+  end
 
+  # GET /lessons/1
   def show
   end
-  
+
+  # GET /lessons/new
   def new
-  	@lesson = Lesson.new
+    @lesson = Lesson.new
   end
 
+  # GET /lessons/1/edit
   def edit
   end
 
+  # POST /lessons
   def create
-  	    lesson = Lesson.new(lesson_params)
-  
+    @lesson = Lesson.new(lesson_params)
+
     if @lesson.save
-      redirect_to my_profile_path, notice: 'Query was successfully created.'
+      redirect_to @lesson, notice: 'Lesson was successfully created.'
     else
       render :new
     end
   end
 
+  # PATCH/PUT /lessons/1
+  def update
+    if @lesson.update(lesson_params)
+      redirect_to @lesson, notice: 'Lesson was successfully updated.'
+    else
+      render :edit
+    end
+  end
+
+  # DELETE /lessons/1
+  def destroy
+    @lesson.destroy
+    redirect_to lessons_url, notice: 'Lesson was successfully destroyed.'
+  end
+
   private
+    # Use callbacks to share common setup or constraints between actions.
+    def set_lesson
+      @lesson = Lesson.find(params[:id])
+    end
 
-  def set_lesson
-  	@lesson = Lesson.find(params[:id])
-  end
-
-  def lesson_params
-  	params.require(:lesson).permit(:paper, :level, :year, :session, :board, :description, :video, :subject_id)
-  end
+    # Only allow a trusted parameter "white list" through.
+    def lesson_params
+      params.require(:lesson).permit(:subject_id, :paper, :level, :year, :session, :board, :description, :video)
+    end
 end
