+++ app/models/lesson.rb
@@ -1,3 +1,4 @@
 class Lesson < ApplicationRecord
   belongs_to :subject
+  mount_uploader :video, VideoUploader
 end
