+++ app/views/lessons/_form.html.erb
@@ -1,0 +1,58 @@
+<%= form_with(model: lesson, local: true) do |form| %>
+  <% if lesson.errors.any? %>
+    <div id="error_explanation">
+      <h2><%= pluralize(lesson.errors.count, "error") %> prohibited this lesson from being saved:</h2>
+
+      <ul>
+        <% lesson.errors.full_messages.each do |message| %>
+          <li><%= message %></li>
+        <% end %>
+      </ul>
+    </div>
+  <% end %>
+
+ <div>
+    <label for="subject_id">Select Subject:</label>
+    <%= form.collection_select :subject_id, Subject.all, :id, :title, prompt: true%>
+  </div>
+
+
+  <div class="field">
+    <%= form.label :paper %>
+    <%= form.number_field :paper %>
+  </div>
+
+  <div class="field">
+    <%= form.label :level %>
+    <%= form.text_field :level %>
+  </div>
+
+  <div class="field">
+    <%= form.label :year %>
+    <%= form.number_field :year %>
+  </div>
+
+  <div class="field">
+    <%= form.label :session %>
+    <%= form.text_field :session %>
+  </div>
+
+  <div class="field">
+    <%= form.label :board %>
+    <%= form.text_field :board %>
+  </div>
+
+  <div class="field">
+    <%= form.label :description %>
+    <%= form.text_area :description %>
+  </div>
+
+  <div class="field">
+    <%= form.label :video %>
+    <%= form.file_field :video %>
+  </div>
+
+  <div class="actions">
+    <%= form.submit %>
+  </div>
+<% end %>
