+++ app/views/lessons/edit.html.erb
@@ -1,0 +1,6 @@
+<h1>Editing Lesson</h1>
+
+<%= render 'form', lesson: @lesson %>
+
+<%= link_to 'Show', @lesson %> |
+<%= link_to 'Back', lessons_path %>
