+++ app/views/lessons/new.html.erb
@@ -1,0 +1,5 @@
+<h1>New Lesson</h1>
+
+<%= render 'form', lesson: @lesson %>
+
+<%= link_to 'Back', lessons_path %>
