+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_08_04_075408) do
+ActiveRecord::Schema.define(version: 2020_08_07_232713) do
 
   create_table "lessons", force: :cascade do |t|
     t.integer "subject_id", null: false
